requirements.txt
fastapi==0.72.0
passlib==1.7.4
PyJWT==2.3.0
pymongo==4.0.1
python-jose==3.3.0
uvicorn==0.17.0
python-multipart==0.0.5